Defer JSON loading until it is needed
Currently we load the JSON file when the middleware object is being
created. If there is a failure in this loading this will cause a 500
error to be raised by the server, without useful logging. This change
defers the loading until it is needed, this way the server will continue
working. Moreover, non-VOMS users can authenticate even if the file
cannot be loaded.