do not fail with 500 if there is SSL data
The refactoring introduced a failure where we failed with a 500 error if
there was no SSL data available in the context, meaning that if a user
requested VOMS authentication without a proxu this will fail (that's OK)
with a 500 (that's not OK). This change raises the proper exception.

(cherry picked from commit 6baa48ce695a927a7ba8c1dcbaa2cc69a0efde93)